pyproject.toml
[project]
name = "tidy-notes"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "openai>=1.93.0",
    "openai-agents>=0.1.0",
    "pydantic>=2.11.7",
    "python-frontmatter>=1.1.0",
]

[dependency-groups]
dev = [
    "pytest>=8.4.1",
]
